--- layers/Schema/packages/categories-wp/src/TypeAPIs/CategoryTypeAPI.php
@@ -1,23 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace PoPSchema\CategoriesWP\TypeAPIs;
-
-use WP_Taxonomy;
-use PoPSchema\Categories\TypeAPIs\CategoryTypeAPIInterface;
-use PoPSchema\TaxonomiesWP\TypeAPIs\TaxonomyTypeAPI;
-
-/**
- * Methods to interact with the Type, to be implemented by the underlying CMS
- */
-class CategoryTypeAPI extends TaxonomyTypeAPI implements CategoryTypeAPIInterface
-{
-    /**
-     * Indicates if the passed object is of type Category
-     */
-    public function isInstanceOfCategoryType(object $object): bool
-    {
-        return ($object instanceof WP_Taxonomy) && $object->hierarchical == true;
-    }
-}
