+++ layers/Schema/packages/categories/src/ComponentContracts/CategoryAPIRequestedContractTrait.php
@@ -4,13 +4,16 @@ declare(strict_types=1);
 
 namespace PoPSchema\Categories\ComponentContracts;
 
+use PoPSchema\Categories\Facades\CategoryTypeAPIFacade;
+use PoPSchema\Categories\TypeAPIs\CategoryTypeAPIInterface;
+
 trait CategoryAPIRequestedContractTrait
 {
-    abstract protected function getTypeAPI(): \PoPSchema\Categories\FunctionAPI;
+    abstract protected function getTypeAPI(): CategoryTypeAPIInterface;
     abstract protected function getTypeResolverClass(): string;
-    protected function getObjectPropertyAPI(): \PoPSchema\Categories\ObjectPropertyResolver
+    protected function getObjectPropertyAPI(): CategoryTypeAPIInterface
     {
-        $cmscategoriesresolver = \PoPSchema\Categories\ObjectPropertyResolverFactory::getInstance();
+        $cmscategoriesresolver = CategoryTypeAPIFacade::getInstance();
         return $cmscategoriesresolver;
     }
 }
