+++ layers/Schema/packages/migrate-categories-wp/migrate/contractimplementations/cms-apis/load.php
@@ -1,5 +1,3 @@
 <?php
 
-require_once 'cms-functionapi.php';
-require_once 'cms-objectpropertyresolver.php';
 require_once 'cms-routingstate-hooks.php';
