--- layers/Schema/packages/migrate-categories/migrate/contracts/cms-apis/cms-objectpropertyresolver-base.php
@@ -1,10 +1,0 @@
-<?php
-namespace PoPSchema\Categories;
-
-abstract class ObjectPropertyResolver_Base implements ObjectPropertyResolver
-{
-    public function __construct()
-    {
-        ObjectPropertyResolverFactory::setInstance($this);
-    }
-}
