--- layers/Schema/packages/migrate-categories/migrate/contracts/cms-apis/cms-objectpropertyresolver-factory.php
@@ -1,17 +1,0 @@
-<?php
-namespace PoPSchema\Categories;
-
-class ObjectPropertyResolverFactory
-{
-    protected static $instance;
-
-    public static function setInstance(ObjectPropertyResolver $instance)
-    {
-        self::$instance = $instance;
-    }
-
-    public static function getInstance(): ObjectPropertyResolver
-    {
-        return self::$instance;
-    }
-}
