--- layers/Schema/packages/migrate-categories/migrate/contracts/cms-apis/cms-objectpropertyresolver-interface.php
@@ -1,10 +1,0 @@
-<?php
-namespace PoPSchema\Categories;
-
-interface ObjectPropertyResolver
-{
-    public function getCategoryID($cat);
-    public function getCategorySlug($cat);
-    public function getCategoryDescription($cat);
-    public function getCategoryCount($cat);
-}
