--- layers/Schema/packages/migrate-categories/migrate/contracts/cms-apis/load.php
@@ -1,7 +1,0 @@
-<?php
-
-require_once 'cms-functionapi-interface.php';
-require_once 'cms-objectpropertyresolver-interface.php';
-require_once 'cms-objectpropertyresolver-factory.php';
-require_once 'cms-objectpropertyresolver-base.php';
-
