+++ layers/Schema/packages/migrate-categories/migrate/contracts/load.php
@@ -1,4 +1,3 @@
 <?php
 
-require_once 'cms-apis/load.php';
 require_once 'vars/load.php';
