+++ layers/Schema/packages/migrate-post-tags/migrate/contracts/cms-apis/cms-functionapi-interface-categories.php
@@ -1,7 +1,9 @@
 <?php
 namespace PoPSchema\PostCategories;
 
-interface FunctionAPI extends \PoPSchema\Categories\FunctionAPI
+use PoPSchema\Categories\TypeAPIs\CategoryTypeAPIInterface;
+
+interface FunctionAPI extends CategoryTypeAPIInterface
 {
 
 }
